Compatibility and security upgrades

+++ EazyStart/src/time/benchmark.c
@@ -102,11 +102,11 @@ void ezs_benchmark_end(const char *name) {
     // 维护 corrected sum of squares [Welford 方差计算]
     // 维护 sumDuration [时间总和]
     const double currentDurationInSeconds = ezs_clock_timespec_to_seconds(duration);
-    const struct timespec previousMean = ezs_clock_timespec_div(entry->sumDuration, (long) (entry->count - 1));
+    const struct timespec previousMean = ezs_clock_timespec_div(entry->sumDuration, entry->count - 1);
     const double previousMeanInSeconds = ezs_clock_timespec_to_seconds(previousMean);
     const struct timespec currentSum = ezs_clock_timespec_add(entry->sumDuration, duration);
     entry->sumDuration = currentSum;
-    const struct timespec currentMean = ezs_clock_timespec_div(currentSum, (long) entry->count);
+    const struct timespec currentMean = ezs_clock_timespec_div(currentSum, entry->count);
     const double currentMeanInSeconds = ezs_clock_timespec_to_seconds(currentMean);
     entry->correctedSumSquaredDuration +=
             (currentDurationInSeconds - previousMeanInSeconds) *
